chore: [M474] make benthic pit observation summaries reusable (for submitted record page)
